refactor: projects using Context
